requirements.txt
Django==4.2.20
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.4.0
Pillow==10.4.0
pytest==8.3.5
pytest-django==4.10.0
coverage==7.5.3
drf-yasg==1.21.9
requests==2.32.3 